requirements.txt
alembic==1.4.3
cffi==1.14.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.2
jwt==1.1.0
Mako==1.1.3
MarkupSafe==1.1.1
Pillow==8.0.1
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.10.1
python-dateutil==2.8.1
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.20
Werkzeug==1.0.1
